requirements.txt
# CompliancePro360 - SaaS Compliance Management System
# Python 3.12+

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2

# Flask Frontend (Production)
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.1

# Database
sqlalchemy==2.0.31
psycopg2-binary==2.9.10
alembic==1.13.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
cryptography==42.0.8

# Task Queue & Scheduler
celery==5.4.0
redis==5.0.8
# celery-beat==2.5.0

# Web Scraping
selenium==4.25.0
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.3.0

# ML/AI
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.2
torch==2.3.1
transformers==4.45.0

# LLM Integration
openai==1.43.0
anthropic==0.36.0
langchain==0.3.5
langchain-community==0.3.3

# Notifications
twilio==9.2.3
sendgrid==6.11.0

# Monitoring & Logging
prometheus-client==0.21.0
python-json-logger==3.2.1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# Utilities
python-dateutil==2.9.0.post0
pytz==2024.2
